Fix multi-containers options
